% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluelink.R
\name{bluelink_dsn}
\alias{bluelink_dsn}
\title{Generate a description of a bluelink file for GDAL}
\usage{
bluelink_dsn(
  x,
  varname = c("atm_flux_diag", "ice_force", "ocean_eta_t", "ocean_force", "ocean_mld",
    "ocean_salt", "ocean_temp", "ocean_tx_trans_int_z", "ocean_ty_trans_int_z",
    "ocean_u", "ocean_v", "ocean_w")
)
}
\arguments{
\item{x}{date string or date(time) format}

\item{varname}{name of variable (that identifies the file)}
}
\value{
string to a thredds file that GDAL can open
}
\description{
Generate a description of a bluelink file for GDAL
}
\examples{
bluelink_dsn()
bluelink_dsn("2020-01-01", varname = "ocean_u")
bluelink_dsn("1993-01-01")
}
