% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluelink.R
\name{read_bluelink}
\alias{read_bluelink}
\title{Title}
\usage{
read_bluelink(
  x,
  varname = c("ocean_salt", "ocean_temp", "ocean_u", "ocean_v", "ocean_w"),
  depth = 1L
)
}
\arguments{
\item{x}{date or datetime object or string}

\item{varname}{variable name one of "ocean_\if{html}{\out{<s>}}" salt, temp, u, v, w  (being salt=salinity, temp=temperature, u,v,w= velocity components in x,y,z direction)}

\item{depth}{depth level (there are 51, from the surface to the bottom) see Details}
}
\value{
SpatRaster
}
\description{
The 'depth' argument is from 1 to 51.
}
\examples{
read_bluelink(varname = "ocean_w")
read_bluelink("2023-01-05", "ocean_salt")
}
