% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluelink.R
\name{read_mld}
\alias{read_mld}
\title{Read Mixed Layer Depth}
\usage{
read_mld(x, ...)
}
\arguments{
\item{x}{date or datetime object or string}

\item{...}{passed to \code{\link[=read_bluelink]{read_bluelink()}}}
}
\value{
SpatRaster
}
\description{
Read Mixed Layer Depth
}
\examples{
a <- read_mld()
b <- read_mld("2023-12-31")
ex <- terra::ext(14, 200, -70, -40)
#terra::crop(a, ex) - terra::crop(b, ex)
}
